rm/urls.py
from shop import urls as shop_urls # <-- Add this at the top
from django.conf.urls import patterns, include, url

from django.contrib import admin
admin.autodiscover()

urlpatterns = patterns('',
    # Examples:
    # url(r'^$', 'rm.views.home', name='home'),
    # url(r'^blog/', include('blog.urls')),

    url(r'^admin/', include(admin.site.urls)),
    (r'^shop/', include(shop_urls)), # <-- That's the important bit
)
