requirements.txt
Django==1.6.5
Pillow==2.4.0
South==0.8.4
django-classy-tags==0.5.1
django-polymorphic==0.5.5
django-shop==0.2.0
jsonfield==0.9.20
pytz==2014.3
wsgiref==0.1.2
